pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hardened-crypto-toolkit"
version = "1.0.0"
description = "A production-grade, hardened CLI toolkit for secure cryptographic operations"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Crypto Toolkit" }]
keywords = ["cryptography", "security", "encryption", "cli", "hashing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]

dependencies = [
    "typer[all]>=0.12.0",
    "cryptography>=42.0.0",
    "argon2-cffi>=23.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.9.0",
    "types-cryptography",
]

[project.scripts]
crypto-toolkit = "crypto_toolkit.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["crypto_toolkit*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["S101"]  # allow assert in tests

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "T20"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=crypto_toolkit --cov-report=term-missing --cov-report=html"
